requirements.txt
scipy==1.11.1
matplotlib==3.7.2
numpy==1.25.1
networkx==3.1
Pillow==10.0.0